Se añadió lo reslacionado con las reseñas

+++ src/main/java/com/info/GamingList/controller/VideojuegoController.java
@@ -26,7 +26,7 @@ public class VideojuegoController {
     }
 
     @GetMapping("/consultar/{id}")
-    public ResponseEntity<?> consultarVideojuego(@PathVariable Long id) {
+    public ResponseEntity<?> consultarVideojuego(@PathVariable int id) {
         try {
             Videojuego videojuego = servicioVideojuego.obtenerVideojuego(id);
             if (videojuego != null) {
